Rename the folder to let the folder more understnadable.
