+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Compose/BottomEditor/AdjustTimeTags/TimeTagRowExtend.cs
@@ -7,7 +7,7 @@ using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Game.Rulesets.Karaoke.Objects;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Compose.BottomEditor.TimeTags
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Compose.BottomEditor.AdjustTimeTags
 {
     public class TimeTagRowExtend : EditRowExtend
     {
